pyproject.toml
[project]
name = "tjm-grounder-task"
version = "0.1.0"
description = "A take-home project for applying to Software Engineer Position at TJM Labs"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "google-genai>=1.60.0",
  "opencv-python-headless>=4.13.0.90",
  "pyautogui>=0.9.54",
  "pygetwindow>=0.0.9",
  "pyperclip>=1.11.0",
  "pyside6>=6.10.2",
  "pytesseract>=0.3.13",
  "python-dotenv>=1.2.1",
  "requests>=2.32.5",
]

[dependency-groups]
dev = ["pre-commit>=4.5.0", "pyright>=1.1.407", "ruff>=0.14.9"]

[tool.setuptools]
package-dir = { "" = "src" }
packages = ['cv_solution', 'llm_solution']

[tool.uv]
package = true

[project.scripts]
start-llm = "llm_solution.gui:run"
start-cv = "cv_solution.gui:run"
notepad-llm = "main:run_llm"
notepad-cv = "main:run_cv"
notepad-cv-first = "main:run_hybrid_cv_first"
notepad-llm-first = "main:run_hybrid_llm_first"

[tool.pyright]
typeCheckingMode = "basic"
reportUnusedImport = "error"
reportUnusedVariable = "error"
reportGeneralTypeIssues = "error"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  'BLE001',
  'C901',
  'D417',
  'E501',
  'E741',
  'ERA001',
  'N802',
  'PLR0913',
  "PLR0912",
  'PLR0915',
  'PLR2004',
  'T201',
]
